+++ src/Logfella/LogWriters/GoogleCloudLogWriter.cs
@@ -2,11 +2,48 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Text.Json;
+using Logfella.Extensions;
+using Microsoft.AspNetCore.Http;
 
 namespace Logfella.LogWriters
 {
     public sealed class GoogleCloudLogWriter : LogWriter
     {
+        public class HttpRequest
+        {
+            public HttpRequest(HttpContext ctx)
+            {
+                var req = ctx.Request;
+
+                RequestMethod = req.Method.ToUpper();
+                RequestUrl = ctx.RequestUrl();
+                RequestSize =
+                    req.ContentLength.HasValue
+                        ? req.ContentLength.Value.ToString()
+                        : "0";
+                UserAgent =
+                    req.Headers.ContainsKey("User-Agent")
+                        ? req.Headers["User-Agent"].ToString()
+                        : "";
+                RemoteIp = ctx.Connection.RemoteIpAddress.ToString();
+                ServerIp = ctx.Connection.LocalIpAddress.ToString();
+                Referer =
+                    req.Headers.ContainsKey("Referer")
+                        ? req.Headers["Referer"].ToString()
+                        : "";
+                Protocol = req.Protocol;
+            }
+
+            public string RequestMethod { get; }
+            public string RequestUrl { get; }
+            public string RequestSize { get; }
+            public string UserAgent { get; }
+            public string RemoteIp { get; }
+            public string ServerIp { get; }
+            public string Referer { get; }
+            public string Protocol { get; }
+        }
+
         public class ErrorContext
         {
             public ErrorContext(Exception ex)
@@ -41,25 +78,26 @@ namespace Logfella.LogWriters
         private readonly string _serviceVersion;
         private readonly bool _useGoogleCloudTimestamp;
         private readonly Dictionary<string, string> _labels;
+        private HttpRequest _httpRequest;
 
         /// <summary>
         /// Initialises a new GoogleCloudLogWriter object.
         /// </summary>
         /// <param name="minSeverity">Minimum severity for the logger to write logs.</param>
-        /// <param name="correlationIdKey">The key under which a correlationId will be logged.</param>
-        /// <param name="correlationId">An ID to correlate a set of logs (e.g. per HTTP request).</param>
         /// <param name="serviceName">Name of the application/service writing logs.</param>
         /// <param name="serviceVersion">Version of the application/version writing logs.</param>
         /// <param name="useGoogleCloudTimestamp">If set to true, then Google Cloud Logging will set its own timestamp for log entries when they have been written to stdout.</param>
         /// <param name="labels">Optional labels to attach to all log entries.</param>
+        /// <param name="correlationIdKey">The key under which a correlationId will be logged.</param>
+        /// <param name="correlationId">An ID to correlate a set of logs (e.g. per HTTP request).</param>
         public GoogleCloudLogWriter(
             Severity minSeverity,
-            string correlationIdKey = "correlationId",
-            string correlationId = "",
             string serviceName = null,
             string serviceVersion = null,
             bool useGoogleCloudTimestamp = false,
-            IDictionary<string, string> labels = null)
+            IDictionary<string, string> labels = null,
+            string correlationIdKey = "correlationId",
+            string correlationId = "")
             : base(
                 minSeverity,
                 correlationIdKey,
@@ -82,6 +120,16 @@ namespace Logfella.LogWriters
             set => _jsonOptions.WriteIndented = value;
         }
 
+        public ILogWriter IncludeHttpRequest(HttpContext ctx)
+        {
+            if (ctx == null)
+                throw new ArgumentNullException(nameof(ctx));
+
+            _httpRequest = new HttpRequest(ctx);
+
+            return this;
+        }
+
         // In the GCP any logs written directly to stdout will be picked up by
         // Google StackDriver Logging and indexed correctly.
         // Structured logs are supported by writing a JSON payload into stdout:
@@ -97,14 +145,6 @@ namespace Logfella.LogWriters
         {
             var msg = new StringBuilder();
             msg.Append(message);
-            if (ex != null)
-            {
-                msg.AppendLine($"");
-                msg.AppendLine($"    Exception Type: {ex.GetType()}");
-                msg.AppendLine($"    Exception Message: {ex.Message}");
-                msg.AppendLine($"    StackTrace:");
-                msg.AppendLine($"        {ex.StackTrace}");
-            }
 
             var logEntry = new Dictionary<string, object>
             {
@@ -122,7 +162,13 @@ namespace Logfella.LogWriters
                 logEntry.Add("serviceContext.version", _serviceVersion);
 
             if (ex != null)
+            {
+                logEntry.Add("@type", "type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent");
                 logEntry.Add("error", new ErrorContext(ex));
+            }
+
+            if (_httpRequest != null)
+                logEntry.Add("httpRequest", _httpRequest);
 
             if (_labels != null && _labels.Count > 0)
                 logEntry.Add("logging.googleapis.com/labels", _labels);